requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
environs==9.5.0
idna==3.3
marshmallow==3.16.0
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
